예외 상황 추가

+++ src/main/java/com/example/leets7th/domain/report/service/ReportService.java
@@ -94,6 +94,11 @@ public class ReportService {
             throw new ReportException(ErrorCode.REPORT_DELETE_NO_PERMISSION);
         }
 
+        //처리된 신고 삭제 불가
+        if(report.getStatus() != ReportStatus.PENDING) {
+            throw new ReportException(ErrorCode.REPORT_ALREADY_HANDLED);
+        }
+
         reportRepository.delete(report);
     }
